+++ Changes.md
@@ -1,6 +1,12 @@
 Changelog for WriteOff.pm
 =========================
 
+v0.33
+-----
+
+- Fixed bug in prelim distr algorithm which resulted in an infinite loop in search of a valid cell to swap with
+- Added `nuke_prelim_round` method to Result::Event, which does what it says on the tin
+
 v0.32 - 18 Feb 2013
 -------------------